+++ src/ElectroMagn/ElectroMagn2D.h
@@ -104,7 +104,7 @@ public:
     void centerMagneticFields();
     
     //! Method used to apply a single-pass binomial filter on currents
-    void binomialCurrentFilter();
+    void binomialCurrentFilter(unsigned int ipass, std::vector<unsigned int> passes);
     
     //! Creates a new field with the right characteristics, depending on the name
     Field *createField( std::string fieldname );
@@ -112,7 +112,6 @@ public:
     //! Method used to compute the total charge density and currents by summing over all species
     void computeTotalRhoJ();
     void addToGlobalRho( int ispec, unsigned int clrw );
-    void computeTotalRhoJs( unsigned int clrw );
     
     //! Method used to compute the total susceptibility by summing over all species
     void computeTotalEnvChi();