+++ src/Params/Params.cpp
@@ -442,7 +442,6 @@ Params::Params( SmileiMPI *smpi, std::vector<std::string> namelistsFiles ) :
     }
 
     // Current filter properties
-    currentFilter_passes = 0;
     int nCurrentFilter = PyTools::nComponents( "CurrentFilter" );
     for( int ifilt = 0; ifilt < nCurrentFilter; ifilt++ ) {
         string model;
@@ -450,7 +449,18 @@ Params::Params( SmileiMPI *smpi, std::vector<std::string> namelistsFiles ) :
         if( model != "binomial" ) {
             ERROR( "Currently, only the `binomial` model is available in CurrentFilter()" );
         }
-        PyTools::extract( "passes", currentFilter_passes, "CurrentFilter", ifilt );
+
+        PyTools::extractV( "passes", currentFilter_passes, "CurrentFilter", ifilt );  //test list
+
+        if( currentFilter_passes.size() == 0 ) {
+            ERROR( "passes cannot be empty" );
+        } else if( currentFilter_passes.size() == 1 ) {
+            while( currentFilter_passes.size() < nDim_field ) {
+                currentFilter_passes.push_back( currentFilter_passes[0] );
+            }
+        } else if( currentFilter_passes.size() != nDim_field ) {
+            ERROR( "passes must be the same size as the number of field dimensions" );
+        }
     }
 
     // Field filter properties
@@ -1017,8 +1027,13 @@ void Params::print_init()
         }
     }
 
-    if( currentFilter_passes > 0 ) {
-        MESSAGE( 1, "Binomial current filtering : "<< currentFilter_passes << " passes" );
+    if (currentFilter_passes.size() > 0){
+        if( *std::max_element(std::begin(currentFilter_passes), std::end(currentFilter_passes)) > 0 ) {
+            for( unsigned int idim=0 ; idim < nDim_field ; idim++ ){
+                std::string strpass = (currentFilter_passes[idim] > 1 ? "passes" : "pass");
+                MESSAGE( 1, "Binomial current filtering : " << currentFilter_passes[idim] << " " << strpass << " along dimension " << idim );
+            }
+        }
     }
     if( Friedman_filter ) {
         MESSAGE( 1, "Friedman field filtering : theta = " << Friedman_theta );