+++ src/Patch/VectorPatch.cpp
@@ -877,6 +877,18 @@ void VectorPatch::sumDensities( Params &params, double time_dual, Timers &timers
             }
         }
     }
+    //Apply boundary conditions for rho and J on axis
+    if ( ( params.geometry == "AMcylindrical" ) && (( *this )( 0 )->vecSpecies.size() > 0) ) {
+        #pragma omp for schedule(runtime)
+        for( unsigned int ipatch=0 ; ipatch<this->size() ; ipatch++ ) {
+            ElectroMagnAM *emAM = static_cast<ElectroMagnAM *>( ( *this )( ipatch )->EMfields );
+            if (emAM->isYmin){
+                for( unsigned int imode = 0 ; imode < emAM->Jl_.size() ; imode++ ) {
+                    ( *this )( ipatch )->vecSpecies[0]->Proj->axisBC( &( *emAM->rho_AM_[imode] )( 0 ), &( *emAM->Jl_[imode] )( 0 ), &( *emAM->Jr_[imode] )( 0 ), &( *emAM->Jt_[imode] )( 0 ), imode );
+                }
+            }
+        }
+    }
     timers.syncDens.update( params.printNow( itime ) );
 } // End sumDensities
 
@@ -936,27 +948,29 @@ void VectorPatch::solveMaxwell( Params &params, SimWindow *simWindow, int itime,
 {
     timers.maxwell.restart();
 
-    for( unsigned int ipassfilter=0 ; ipassfilter<params.currentFilter_passes ; ipassfilter++ ) {
-        #pragma omp for schedule(static)
-        for( unsigned int ipatch=0 ; ipatch<this->size() ; ipatch++ ) {
-            // Current spatial filtering
-            ( *this )( ipatch )->EMfields->binomialCurrentFilter();
-        }
-        if (params.geometry != "AMcylindrical"){
-            SyncVectorPatch::exchangeAlongAllDirections<double,Field>( listJx_, *this, smpi );
-            SyncVectorPatch::finalizeExchangeAlongAllDirections( listJx_, *this );
-            SyncVectorPatch::exchangeAlongAllDirections<double,Field>( listJy_, *this, smpi );
-            SyncVectorPatch::finalizeExchangeAlongAllDirections( listJy_, *this );
-            SyncVectorPatch::exchangeAlongAllDirections<double,Field>( listJz_, *this, smpi );
-            SyncVectorPatch::finalizeExchangeAlongAllDirections( listJz_, *this );
-        } else {
-            for (unsigned int imode=0 ; imode < params.nmodes; imode++) {
-                SyncVectorPatch::exchangeAlongAllDirections<complex<double>,cField>( listJl_[imode], *this, smpi );
-                SyncVectorPatch::finalizeExchangeAlongAllDirections( listJl_[imode], *this );
-                SyncVectorPatch::exchangeAlongAllDirections<complex<double>,cField>( listJr_[imode], *this, smpi );
-                SyncVectorPatch::finalizeExchangeAlongAllDirections( listJr_[imode], *this );
-                SyncVectorPatch::exchangeAlongAllDirections<complex<double>,cField>( listJt_[imode], *this, smpi );
-                SyncVectorPatch::finalizeExchangeAlongAllDirections( listJt_[imode], *this );
+    if (params.currentFilter_passes.size() > 0){
+        for( unsigned int ipassfilter=0 ; ipassfilter<*std::max_element(std::begin(params.currentFilter_passes), std::end(params.currentFilter_passes)) ; ipassfilter++ ) {
+            #pragma omp for schedule(static)
+            for( unsigned int ipatch=0 ; ipatch<this->size() ; ipatch++ ) {
+                // Current spatial filtering
+                ( *this )( ipatch )->EMfields->binomialCurrentFilter(ipassfilter, params.currentFilter_passes);
+            }
+            if (params.geometry != "AMcylindrical"){
+                SyncVectorPatch::exchangeAlongAllDirections<double,Field>( listJx_, *this, smpi );
+                SyncVectorPatch::finalizeExchangeAlongAllDirections( listJx_, *this );
+                SyncVectorPatch::exchangeAlongAllDirections<double,Field>( listJy_, *this, smpi );
+                SyncVectorPatch::finalizeExchangeAlongAllDirections( listJy_, *this );
+                SyncVectorPatch::exchangeAlongAllDirections<double,Field>( listJz_, *this, smpi );
+                SyncVectorPatch::finalizeExchangeAlongAllDirections( listJz_, *this );
+            } else {
+                for (unsigned int imode=0 ; imode < params.nmodes; imode++) {
+                    SyncVectorPatch::exchangeAlongAllDirections<complex<double>,cField>( listJl_[imode], *this, smpi );
+                    SyncVectorPatch::finalizeExchangeAlongAllDirections( listJl_[imode], *this );
+                    SyncVectorPatch::exchangeAlongAllDirections<complex<double>,cField>( listJr_[imode], *this, smpi );
+                    SyncVectorPatch::finalizeExchangeAlongAllDirections( listJr_[imode], *this );
+                    SyncVectorPatch::exchangeAlongAllDirections<complex<double>,cField>( listJt_[imode], *this, smpi );
+                    SyncVectorPatch::finalizeExchangeAlongAllDirections( listJt_[imode], *this );
+                }
             }
         }
     }
@@ -1430,9 +1444,6 @@ void VectorPatch::solvePoisson( Params &params, SmileiMPI *smpi )
         //The MPI rank owning it is
         int rank_XmaxYmin = smpi->hrank( patch_YminXmax );
 
-
-        //cout << patch_YmaxXmin << " " << rank_XminYmax << " " << patch_YminXmax << " " << rank_XmaxYmin << endl;
-
         if( smpi->getRank() == rank_XminYmax ) {
             Ex_XminYmax = ( *this )( patch_YmaxXmin-( this->refHindex_ ) )->EMfields->getEx_XminYmax();
             Ey_XminYmax = ( *this )( patch_YmaxXmin-( this->refHindex_ ) )->EMfields->getEy_XminYmax();
@@ -1544,11 +1555,9 @@ void VectorPatch::solvePoissonAM( Params &params, SmileiMPI *smpi )
 
     for( unsigned int ipatch=0 ; ipatch<this->size() ; ipatch++ ) {
         ( *this )( ipatch )->EMfields->initPoisson( ( *this )( ipatch ) );
-        //cout << std::scientific << "rnew_dot_rnew_local = " << rnew_dot_rnew_local << endl;
         ElectroMagnAM *emAM = static_cast<ElectroMagnAM *>( ( *this )( ipatch )->EMfields );
         emAM->initPoissonFields( ( *this )( ipatch ) );
     }
-    // //cout << std::scientific << "rnew_dot_rnew_local = " << rnew_dot_rnew_local << endl;
 
     std::vector<Field *> El_;
     std::vector<Field *> Er_;
@@ -1607,7 +1616,6 @@ void VectorPatch::solvePoissonAM( Params &params, SmileiMPI *smpi )
         }
         
         iteration = 0;//MESSAGE("Initial error parameter (must be 1) : "<<ctrl);
-        //cout << std::scientific << ctrl << "\t" << error_max << "\t" << iteration << "\t" << iteration_max << endl;
         while( ( ctrl > error_max ) && ( iteration<iteration_max ) ) {
             iteration++;
         
@@ -1855,10 +1863,8 @@ void VectorPatch::solveRelativisticPoisson( Params &params, SmileiMPI *smpi, dou
     for( unsigned int ipatch=0 ; ipatch<this->size() ; ipatch++ ) {
         ( *this )( ipatch )->EMfields->initPoisson( ( *this )( ipatch ) );
         rnew_dot_rnew_local += ( *this )( ipatch )->EMfields->compute_r();
-        //cout << std::scientific << "rnew_dot_rnew_local = " << rnew_dot_rnew_local << endl;
         ( *this )( ipatch )->EMfields->initRelativisticPoissonFields( ( *this )( ipatch ) );
     }
-    //cout << std::scientific << "rnew_dot_rnew_local = " << rnew_dot_rnew_local << endl;
     MPI_Allreduce( &rnew_dot_rnew_local, &rnew_dot_rnew, 1, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD );
 
     std::vector<Field *> Ex_;
@@ -1935,7 +1941,6 @@ void VectorPatch::solveRelativisticPoisson( Params &params, SmileiMPI *smpi, dou
     if( smpi->isMaster() ) {
         DEBUG( "Starting iterative loop for CG method" );
     }
-    //cout << std::scientific << ctrl << "\t" << error_max << "\t" << iteration << "\t" << iteration_max << endl;
     while( ( ctrl > error_max ) && ( iteration<iteration_max ) ) {
         iteration++;
 
@@ -2069,9 +2074,6 @@ void VectorPatch::solveRelativisticPoisson( Params &params, SmileiMPI *smpi, dou
         int patch_YminXmax = domain_decomposition_->getDomainId( xcall );
         //The MPI rank owning it is
         int rank_XmaxYmin = smpi->hrank( patch_YminXmax );
-
-
-        //cout << patch_YmaxXmin << " " << rank_XminYmax << " " << patch_YminXmax << " " << rank_XmaxYmin << endl;
 
         if( smpi->getRank() == rank_XminYmax ) {
             Ex_XminYmax = ( *this )( patch_YmaxXmin-( this->refHindex_ ) )->EMfields->getExrel_XminYmax();
@@ -2289,10 +2291,8 @@ void VectorPatch::solveRelativisticPoissonAM( Params &params, SmileiMPI *smpi, d
 
     for( unsigned int ipatch=0 ; ipatch<this->size() ; ipatch++ ) {
         ( *this )( ipatch )->EMfields->initPoisson( ( *this )( ipatch ) );
-        //cout << std::scientific << "rnew_dot_rnew_local = " << rnew_dot_rnew_local << endl;
         ( *this )( ipatch )->EMfields->initRelativisticPoissonFields( ( *this )( ipatch ) );
     }
-    // //cout << std::scientific << "rnew_dot_rnew_local = " << rnew_dot_rnew_local << endl;
 
     std::vector<Field *> El_;
     std::vector<Field *> Er_;
@@ -2382,7 +2382,6 @@ void VectorPatch::solveRelativisticPoissonAM( Params &params, SmileiMPI *smpi, d
         }
         
         iteration = 0;//MESSAGE("Initial error parameter (must be 1) : "<<ctrl);
-        //cout << std::scientific << ctrl << "\t" << error_max << "\t" << iteration << "\t" << iteration_max << endl;
         while( ( ctrl > error_max ) && ( iteration<iteration_max ) ) {
             iteration++;